requirements.txt
# Real-Time Multi-API Data Dashboard - Dependencies

# Core
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# API Clients
requests>=2.31.0
yfinance>=0.2.28
alpha-vantage>=2.3.1

# Data Processing (PySpark optional - uncomment for large-scale data)
# pyspark>=3.5.0

# Visualization & Dashboard
streamlit>=1.28.0
streamlit-autorefresh>=0.0.6
plotly>=5.18.0

# Machine Learning (optional)
scikit-learn>=1.3.0

# Utilities
schedule>=1.2.0
